+++ R/steps.R
@@ -92,9 +92,12 @@ Cicerone <- R6::R6Class(
 #' @param tab The name of the tab to set.
 #' @param is_id Whether the selector passed to `el` is an HTML id, set to `FALSE` to use
 #' other selectors, e.g.: `.class`.
+#' @param on_highlighted An arbitrary piece of JavaScript code to run when the step is highlighted,
+#' generally a callback function. This is effectively a string that is evaluated JavaScript-side.
     step = function(el, title = NULL, description = NULL, position = NULL, 
       class = NULL, show_btns = NULL, close_btn_text = NULL,
-      next_btn_text = NULL, prev_btn_text = NULL, tab = NULL, tab_id = NULL, is_id = TRUE) {
+      next_btn_text = NULL, prev_btn_text = NULL, tab = NULL, tab_id = NULL, is_id = TRUE,
+      on_highlighted = NULL) {
 
       assertthat::assert_that(!missing(el), msg = "Must pass `el`.")
 
@@ -115,6 +118,8 @@ Cicerone <- R6::R6Class(
       if(!is.null(prev_btn_text)) popover$prevBtnText <- prev_btn_text
 
       step = list(element = el, tab_id = tab_id, tab = tab)
+
+      if(!is.null(on_highlighted)) step$onHighlighted <- on_highlighted
 
       if(length(popover)) step$popover <- popover