+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Add `tab` and `tab_id` arguments to trigger open tabs.
 - Add `is_id` argument to `step` and `highlight` method to allow using other selectors than `#id`, see [#7](https://github.com/JohnCoene/cicerone/issues/7)
+- Add `on_highlighted` to run JavaScript functions when the step is highlighted, see [#13](https://github.com/JohnCoene/cicerone/issues/13)
 
 # cicerone 1.0.2